bundles the loading overlay into Container uikit component
