+++ client/components/pages/submission-system/create-program/index.js
@@ -20,9 +20,6 @@ import CREATE_PROGRAM_MUTATION from './CREATE_PROGRAM_MUTATION.gql';
 import createProgramSchema from './validation';
 import { useMutation } from 'react-apollo-hooks';
 import useCommonToasters from 'components/useCommonToasters';
-import DnaLoader from 'uikit/DnaLoader';
-import useTheme from 'uikit/utils/useTheme';
-import color from 'color';
 
 const SectionTitle = styled('h3')`
   ${({ theme }) => css(theme.typography.subtitle2)};
@@ -53,30 +50,6 @@ const createProgramInput = formData => ({
   cancerTypes: formData.cancerTypes,
   primarySites: formData.primarySites,
 });
-
-const LoadingOverlay = () => {
-  const theme = useTheme();
-  return (
-    <div
-      css={css`
-        position: absolute;
-        left: 0px;
-        right: 0px;
-        top: 0px;
-        bottom: 0px;
-        background: ${color(theme.colors.white)
-          .alpha(0.7)
-          .hsl()
-          .string()};
-        display: flex;
-        justify-content: center;
-        align-items: center;
-      `}
-    >
-      <DnaLoader />
-    </div>
-  );
-};
 
 export default () => {
   const toaster = useToaster();
@@ -126,12 +99,11 @@ export default () => {
       }
     >
       <Container
+        loading={formDisabled}
         css={css`
           margin: 10px auto;
           padding: 10px 40px;
           max-width: 875px;
-          position: relative;
-          overflow: hidden;
         `}
       >
         <ProgramForm
@@ -142,7 +114,6 @@ export default () => {
             </Link>
           }
         />
-        {formDisabled && <LoadingOverlay />}
       </Container>
     </SubmissionLayout>
   );
